backend/pyproject.toml
[project]
name = "dnd-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "pymongo>=4.10.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
